Simplify storyboard example & style fix
