+++ Example/Demo.storyboard
@@ -6,10 +6,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Storyboard Load More View Controller-->
+        <!--Storyboard View Controller-->
         <scene sceneID="oB3-c8-g1U">
             <objects>
-                <viewController storyboardIdentifier="demo" id="5n3-mX-NaJ" customClass="StoryboardLoadMoreViewController" customModule="IGListKitExamples" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="demo" id="5n3-mX-NaJ" customClass="StoryboardViewController" customModule="IGListKitExamples" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="VEL-Rj-Wcy"/>
                         <viewControllerLayoutGuide type="bottom" id="wYR-29-GPY"/>
