+++ Tests/Objects/IGTestStoryboardCell.m
@@ -11,4 +11,10 @@
 
 @implementation IGTestStoryboardCell
 
+- (id)initWithCoder:(NSCoder *)aDecoder {
+    self = [super initWithCoder:aDecoder];
+    self.label.text = @"";
+    return self;
+}
+
 @end
